Replaced old address data w/ new (already inserted to table).
Added copy of SQLBase to data_files pkg.

Added SQLDataInserter.java which (as is) would attempt to add all
elements from the FAKE_ADDRESS.csv file, into the SQL db defined in
SQLBase, specifically to the schema provided in SQLDataInserter's
creation of the connection.

SQLDataInserter should be easily adaptable to any of the other .csv(s)
by simply changing the file name, quantity of question marks, rowCount
(which could be replaced by a hasNext, like how scanner2 processes each
column), and changing the quantity of calls to provide the prepared
statement with the data gathered from the scanner, among ensuring the
respective functions are called respective to the type of the data being
grabbed.